Add dropdown to select academic year
This adds logic to:

1) Pull the current academic year from the metadata
2) Provide a dropdown of academic years to change the data

The code for parsing the semesters into an academic year list does feel
a bit clunkhy. I think we may want to look at changing the semesters
field in the metadata to be a list of semester objects with term,
calendarYear, and academicYear values, so that we can better format the
data based on what we need.