Refactor debouncing logic
Running the complete filtering logic within the setTimeout block was
introducing some additional delays when first loading the page and when
the modals closed, i.e. when the list of courses updates but the filters
don't. The best way for fixing that I could find was to create this
second set of filters, and change the debouncing logic so that it
instead moves the current filters (the "value" of the input) to the
"activeFilters", and only then does it recalculate the memoized list.

Since the useMemo hook will update when either the currentCourses or the
activeFilters value change, we can now immediately update our list of
courses after loading the page or saving a course.

(Side note: Subjectively speaking, it feels like the delay I'm seeing is
more than 100ms, which I think might be a side effect of how react's
reconciliation algorithms work, where having any additional delay is
pushing updates further down the call stack. That might be worth further
investigation though.)

This also removes the shouldClose*Modal state fields and reverts to
calling the close*Modal functions directly in the onSave hooks, since
removing that 100ms delay seems to put UI updates back in order and teh
modals are now closing correctly before the list updates, so we should
not longer get the stale data appearing after saving. Additionally,
those extra state values seem to have been triggering some extra
re-renders in certain cases -- I think the problem was that updating
shouldClose*Modal inside the useEffect hook was triggering the effect
again immediately after, and causing some unintended recomputation.