Add Show Retired Checkbox
This replaces the hardcoded 'showRetired' value with a showRetired state. The checkbox toggles the showRetired value. A flex wrapper, MenuFlex, was added to control the formatting of the options above the Course Instance table. As more items are added within this flex component, the items will be formatted with equal spacing between each item.