Generate API endpoint tests
Rather than duplicate the tests for each user for each endpoint (80+
tests), we can create two arrays and run a double-for loop to
automatically generate all of tests. This does mean adding a little
extra logic to determine which tests we should run, but it has the
advantage of consolidating all of the data about our endpoints into a
single file for reference.

Note that the only thing we're checking here is the
authentication/authorization layer, and we're doing so by looking at the
status code returned from the endpoint to see if it is/isn't 403 or 401.
This means we're letting some other errors (data validation, typeorm
errors, etc) bubble up, since that's all taking place behind the auth
layer.

There was a little weirdness with the current user endpoint, as that's
the only one that returns a 401 (UNAUTHORIZED) instead of a 403
(FORBIDDEN). Since that's the only one that requires authentication
(i.e. no anonymous access) but doesn't require a specific group, we can
use those conditions to test for the 401 instead.