Change element await
The loading indicator being gone doesn't necessarily mean that the data
has loaded into the table.

As a result, the loading indicator would disappear (as dictated by the
`waitForElementTobeRemoved` call), but the data still hadn't been loaded
into the table.

Waiting for the actual table data to appear produces a more
deterministic result