Implement group checking in custom hook
This is similar enough to our server-side GroupGuard to merit the
similar naming, but it works in a slightly different way.

On the server, we can apply the guard to a single endpoint or subpath,
and we know that we'll only need to apply one group level in each guard
call. I originally started creating this hook in the same way, where you
could pass in a group and get back a simple boolean.

However, everywhere that I planned to use it in the client actually
needed to know information about multiple group memberships (e.g. the
header dynamically changing the menu items). So I wound up pivoting to
this version, where get back a dictionary of permission-to-boolean
mappings that we can descructure to get our different permission levels.