Dynamically render edit buttons for admins
Only admins can edit data in the table, so we only need to render the
edit buttons if the current user is one.

I originally thought that the hook call needed to be at the top
CoursesPage level, with the isAdmin value passed down through each
layer, because of the errors we've had in the past about calling hooks
conditionally or inside of other hooks. But since we've split the
CourseInstanceTableBody out into its own component and we're using
React.memo instead of useMemo, this doesn't seem to be an issue here.