Dynamically show log in/out button
Mirroring the behavior of the current app, the log-in button will appear
on the public-facing site and clicking it will bump the user over to the
private site (assuming they are allowed to access it).

On the private site, the button will just log the user out of
HarvardKey, which is what it did previously.