Show 401/403 pages based on permissions
Technically, a 404 page implies that the page doesn't exist, so it makes
more sense to show errors that the user isn't allowed to view a page, or
needs to log in to view a page.

Note that this will need to be refactored when we implement the
public-facing site, since there will actually be pages that can be
viewed while not logged in.