Convert AppRouter tests to e2e
I had initially tried to extend our initial AppRouter tests, but I
eventually hit a wall where the client needed actual data to render in
the tables, and stubbing out all the API endpoints would have quickly
hit a point of diminishing returns. This instead moves the existng
routing tests into our end-to-end style, and adds additional tests for
our 401/403 conditions.

As noted in @TODO comments, we'll need to change some of these tests
when we have a public-facing site to test.