Update the controller and service to utilize the CourseInstanceUpdateDTO
Instead of updating the entire CourseInstanceResponseDTO, we are now only sending the course instance id and the CourseInstanceUpdateDTO, which includes the offered and enrollment values. I found that we do not need the entire CourseInstanceResponseDTO as the return type in order to populate the local course value in the course instance table body, so I set up the response to return the updated offered and enrollment values.