Do a better job fixing URL handling
I had misread the docs for the URL class in my previous work, and
passing the `/login` path in the constructor was overwriting any
existing pathnames, i.e the `/course-planner/` segment of our URL.

However, concatenating onto the existing pathname also has a pitfall
since it can sometimes end in a `/` and sometimes not, resulting in a
`//` double slash appearing in the stringified version. This requires
adding another check to see whether we need to include the leading slash
or not when appending.