Add Courses report endpoint
This endpoint uses the streaming I/O methods of ExcelJS to write the
spreadsheet data to the response stream as each course is processed, so
it should start the report download faster with lower memory overhead;
the latter of which could be a factor when we run the app in ECS.

Right now it presents a report starting with the current academic year
and running to the last year in the database. These are computed inteh
controller, so they could be replaced with query parameters when we want
to allow users to define their own report range.

The xlsx output doens't include any styling or other features, and omits
the top line from the previous report, which used merged cells to show a
Semester header.