Add "Download Course Report" button
I waffled over whether this should be a button, or if it should really
be a link to the download url. Semantically a link makes more sense, but
stylistically it should look like a button so we'll go with that for
now.

Of courses this may change in the near future, if I waffle back the
other way.