Revert conditional rendering of Offered dropdown values
Spoke to C Rogers and decided that it would be best not to do conditional rendering of offered options until we speak to stakeholders about how we should display values of previous years (we need to decide whether or not we will even have an edit button for past academic year offered fields). Also, another point to think about is how past retired courses will render if there is no longer a retired option in the dropdown.