Fix handling of BadRequestExceptions
When there was only one error (testing the case of trying to retire a course instance from a previous academic year), there was an error indicating that we could not map over the serverError.message, since this was not an array (it was just a string). Instead, we are checking to see if serverError.message is an array (it is an array when there is more than one message) and mapping over it to set the error message instead of setting the error message directly.