Download report asynchronously
One of the drawbacks of the navigation-based approach set up in the
previous commit is that errors from the server would pull users out of
the app and, most likely, show them a raw JSON error. This is not ideal.

The best solution I found for this is to use fetch() to get the binary
data in the background, convert it into a data URL, then create a
temporary link to that data URL that we click once and immediately
remove. It's a little hacky, but it gives us some clear advantages:

1. Handle the intermediate "downloading" state in the app, by disabling
   the button and changing its message
2. Handle any errors within the app, via our global message
3. Avoid a case I was running into on Firefox, where clicking the button
   opens a new tab
4. Is semantically consistent, since we're using a button to start an
   asynchronous action, not to navigate.

I've extacted the downloading logic into its own util function, since
I'm expecing we'll use this for the Faculty Report as well.