Update service to handle retired values
Course instances can be retired if they are from the current or for future academic years. If a course instance is marked as retired, the instances that follow should also be marked as retired. If a course is in the same academic year or later and it is marked as retired, if the user changes the offered value to something other than retired, the instance itself will be updated to what the user saved it as, and the future academic years will be updated to OFFERED.BLANK. There is also a check that makes sure that the user isn't changing from OFFERED.RETIRED to OFFERED.RETIRED - in this case, I think it makes sense not to overwrite the offered values of the later instances, which could have changed to other OFFERED values since we can alter future offered instance values even after retiring an instance.