Add end to end tests to check functionality of retiring and un-retiring courses
These tests check the effects of setting courses to Retired and setting retired courses to other values. When you set a course to retired, if it is a valid request (you can only set current and future academic year instances to reitred), later instances that exist will be retired. If you un-retire a course (we are only allowed to un-retire courses of the current or future academic years), it should update that instance to the requested value and update the offered value for future instances to OFFERED.BLANK.