Separate & Improve formatting logic
I realized that it is possible to have a meeting day/time without an
assigned room, and adding that in made it clear that having this logic
defined twice in the spring/fall map functions was going to be far more
complicated than necessary.

I also thought it was a little odd to use a comma to separate names in
the "lastName, firstName" format, so I further split that join argument
off into its own value to make it easier to update each entry.