Test basic spreadsheet output
Per discussion in our spring planning meeting, the spreadsheet does not
need to be deeply tested, but this minimal test suite makes sure that
the output has the expected number of rows, and includes teh expected
range of years. We may opt to build out this formatting in the future,
if/when we add the additonal styling to the sheet.