Inject HTTPS_* environment variables
The trio of environment variables will be used to optionally configure
Nest to serve https responses. This is primarily for our production
deployment, where we'll want to make sure that traffic between our load
balancer and our ECS container is encrypted.

As an implementation note: The load balancer will handle the SSL
termination from our clients, and will repond with a valid certificate
verifying its hostname; this container<=>LB traffic does not need to
handle hostname verification, just the encryption of traffic.