Update e2e tests to wait for elements before proceeding with tests
After changing the year dropdown, this waits for the loading spinner to go away before trying to look for the table rows. This also updates the test to wait after checking the 'show retired' checkbox to find the course number before proceeding with the tests. This seems to be fixing an issue in which these last couple tests fail due to not being able to find the row element.