Fix key formatting for AWS Secrets
When the certificate/key are created in the AWS secrets manager, the
newline character are all replaced with spaces. This causes an error, as
the underlying TLS library expects at least the header and footer to be
on their own lines. Unfortunately it's not quite as easy as just
replacing all spaces with newlines, as the header and footer contain
spaces on their own. Instead, we'll include the lookahead to make sure
that the full header and footer are presevered.

Note that when I generated a key locally I got a header of:

-----BEGIN RSA PRIVATE KEY-----

But in other cases -- such as the key Fisher generated for AWS, it's:

-----BEGIN PRIVATE KEY-----

I've included the "RSA" in the alternator for widest compatiblity, since
it will still work even when RSA is not in the string.