Strip trailing slash from pathname
The value of pathname returned from the URL could be /course-planner or
/course-planner/, depending on what's set as the environment variable.
This will remove a trailing slash ensuring that:

1. If there is no trailing slash to begin with, none will be returned
2. If there is a tailing slash, it will be removed
3. If there is no pathname, the default of "/" will still be removed

Since we're passing the pathname to the global prefix option, having a
trailing slash, or just "/", could result in a double-slash (//) in the
path. This is usually not a problem, but could cause some unexpected
behavior in the app routing.