Don't disable built-in logger
With this set to false, Nest would not log any message until our custom
LogService became available. While this makes our logs more consistent,
it also means that when there's an error that prevents the LogService
from initializing on startup, we don't get any log output.

Removing this setting means that Nest will use it's built-in logger
until the custom one becomes available, then switch to it. That way, we
can at least see when there are errors and address them.

When we do update Nest to version 8[1], there's a new option called
`bufferLogs`, which will hold the logs until the custom LogService is
created, so we can try using that and see if it helps. (It looks like
there are a number of changes[2] to the Logging configuration that were
merged in v8, so we'll likely need to refactor much of that setup
regardless)

[1]: https://github.com/nestjs/nest/releases/tag/v8.0.0
[2]: https://github.com/nestjs/nest/pull/6221