Add semester service to add a new academic year
This function checks whether the current month is June. If it is, it will check if the new academic year exists before trying to create it. The semester repository is queried so that we can get the existing data for absences and non class events. The reason why we are not querying for the course instances via the semester repository is because we need to know the spring offered values for each instance in order to know whether to mark the new course instances as OFFERED.RETIRED or OFFERED.BLANK. If the spring semester of the last existing academic year is retired, the new instances should also be retired.