Replace access to course instance service
Instead of calling the course instance service to get the course instance data, we are accessing the course instances via the findOneOrFail call to the semester repository. By creating a mapping between the course id and the spring instances, we can access the spring offered data while processing the fall data. (We need the spring offered data in order to determine the correct offered value for the newly created fall semester).