Add semester service tests that check that the mockSemesterRepository methods are called the appropriate number of times
These tests check the behavior of the semester service before, on, and after June 1st. We are checking to see the number of times that findOneOrFail and save methods of the mockSemesterRepository are called in order to see whether the service attempts to make new instances depending on the time of year.