Fix docker publication version tagging
The previous code was always building and pushing the "main" branch, but
tagging it with the "version", "stable", and "qa" tags, regardless of
where the merge occurred. This should address the problem by always
using the tagged url as the context for the build. It also attempts to
simplify the logic of tagging witha shell script that assembles a
comma-separated list of tags, instead of trying to use the "if" option
to the job.