Add client publication stage
Adds a stage that generates two front-end bundles for the
course-planner, with the urls for the prod and qa environments set in
the repository secrets. Each bundle is published as a tarball to the
releases section in github.

I originally started doing this as a completely separate workflow, but I
felt there was enough crossover that it made more sense to combine them
in a single workflow and move the common steps into a pre-publish stage.