Replace getELementById call with a stable ref
The previous implementation used `document.getElementById` instead of a
stable ref. This in turn was because the ref for the notes button was
not being populated in the `refTable` object. This in turn was not
happening because `buttonRef` had not been passed down from
the `FieldContent` props to the add/edit notes button itself.

Refs are really just functions, so the act of using the ref, causes the
`useRef` hook to be called, which in turn populates the `refTable`
object with a ref to that component.

This last crucial step of passing the ref in to the button had been
overlooked meaning that the call to
`refTable.current[modalButtonId].focus()` then raised an error because
there was no ref to call `.focus()` on.

More on how `useRef` and `refTable` works available in
5e6122d6cdc734a3a2c8a46efa706ec1b6743b06