Add newAcademicYear parameter to addAddAcademicYear function
This update adds a newAcademicYear parameter to allow us more flexibility in creating years. The logic of the current date and when to run this function will be separately handled outside of this function. This change allows us to more easily shift to a 5 year plan (vs the current 4 year plan). An exception is thrown if the academic year previous to the requested new academic year does not yet exist - it's important for this preceding year to exist, because the new year is based off of the last existing year (this is especially important in terms of the "offered" values)