Remove top-level ternary around modal
The ternary was used to block the modal from rendering if
`modalData.visible` was false. However, this is a little redundant as
the modal itself accepts an `isVisible` prop which is used to control
visibility.

Removal of the ternary then caused problems with properties of `course`
as `course` did not yet have a value when the modal was first rendered.
To fix _this_ references to properties of `course` were changed to use
the optional chaining prop, and the `courseNotes` state field was
initialized to an empty string(before being set to the value of
`course.notes` after initial render).