Wrap assertions in setTimeout call to allow render cycle to complete
The render cycle was not previously complete when calls to `setTimeout`
were made. This meant that the modal was being closed, and then the
assertion was immediately being run before the DOM had time to
re-render.

This then meant that the test failed prematurely before the render cycle
was complete.

The setTimeout function was used around the assertion to address this by
waiting for the render cycle to fully complete before trying to assert
anything