Use getByText instead of getByRole
Usage of `getByRole` can be slow in some instances due to the way the
underlying library searches for element roles. Instead, using
`getByText` is faster.

In the case of the latter example on line 64, we don't need to use
`.parentElement` like we did on line 37 because we don't actually care
about a reference to the modal, we just want to wait for the modal
header to disappear(meaning that by implication - the modal too has
disappeared)