Add AbsencePopulationService
This service populates absences for our test database. This is needed in order to test that adding a new academic year also adds absences for each faculty with a value of ABSENCE_TYPE.PRESENT for every faculty member. Since there was no absence test data previously, there wasn't a way to test that the new absences are created with the correct absence type. There is an absence created for each faculty for each semester.