Test the file download process
These tests are weird because this whole process for downloading the
file is weird, and is relying on a couple of (currently) browser only
APIs. So it's annoyingly close to testing implementation details, but
should at least confirm that our fetch is being executed, and then that
things are happening with the data we get back.

Note: I did need to minorly tweak the downloadAttachment function to
make sure that the hidden download link would actually be testable.