Add integration tests to check addAcademicYear
These tests check the effect of creating a new academic year on the created course instances, faculty absences, and non class events. When the previous existing spring semester's course instance is retired, the new course instance is also marked as retired. It also checks that a bad request exception is thrown if the year previous to the academic year the user is trying to create does not yet exist.