Add OnApplicationBootstrap to SemesterService
OnApplicationBootstrap enables us to run the addAcademicYear function when the app starts. A check was added to see whether we are in the month of June before running the function, and there is a check within the function to only add the new academic year only if it doesn't already exist.