Split out behavior into describe/context blocks
Trying to re-render within the same describe block was causing the
results of one render call to pollute the other meaning that when the
component was rendered with a different user "logged in" the tests would
occasionally see the result from the top-level `render()` call which
would cause the test to fail (or pass) when it shouldn't.

The fix for this was to separate out the different cases into
describe/context blocks so that each `render()` call with different
users was scoped to that block.