Generate key/cert for redis container
Background Context: We need to run the connection between our containers
and redis over TLS in production. Previously we were setting this
conditionally based on the NODE_ENV value, but this creates new problems
if we need to run under a different NODE_ENV value while testing (e.g.,
using 'development' to bypass Harvard Key while we make sure everything
is running correctly.

Rather than introduce another flag to run in TLS mode conditionally, I
think it makes sense to run our local containers on TLS as well, so that
our configuration better matches our production environment, as we did
with our postgres config.

Unfortunately setting up TLS is a little more complicated with the redis
container, as there is no default init script that we can run. Instead,
we're overriding the "CMD" option (via docker-compose) to launch a
custom init script that generates the keys necessary and sets the
environment values needed to configure TLS, then executes the setup.sh
and run.sh scripts used by the container normally. This feels a little
fragile, but it's the best way I can find to create the key/cert on
startup.