Modify Redis config options
In order to connect to the local redis container with its self-signed
certificate, we need to set the tls.rejectUnauthorized option to false in
our Redis client config. This cannot be done as a URL parameter,
unfortunately, so we need to change our redis config option to return an
connection options object instead.