Move column selection logic into ViewModal
As we're no longer updating the columns "live", it doesn't make as much
sense to have the SemesterTable defined as the children prop to the
ViewModal, and we can instead create a separate state object inside the
modal that tracks which columns the user wants. We can then handle
persisting those columns in SessionStorage and updating the effective
set of columns in the onClose handler.

This also makes a few minor changes to where the ViewModal component is
located inside the page, and how the Ref is managed, to bring the
ViewModal more in line with how our other modals have been set up.