Initial tests of useStoredState behavior
I don't love these tests, since there's a lot of setup and stubbing
necessary to make them work. But I think it's important to have some
isolated tests of the hook behavior. It also presents a blueprint for
how to implement a fake Storage backend that can be used in the next
set of end-to-end tests.