End-to-end tests of saved custom columns
This again relies on a shimmed out version of window.sessionStorage,
which is not ideal but is probably the best we can do here. I've really
only implemented two "tests", one with session stubbed and on without,
though each incorporates a number of assertions (with identifying error
messages to check the behavior as the "user" moves through the flow.