Replace BadRequestException with Error and fix associated tests
Since a bad request exception implies that there was something wrong with the request itself, it is better to use the more general Error instead here. Since the error does not have a message property, the error was "stringify"'ed in order to check that the error contains the correct error text.