Update logging for adding academic years
Different levels of logging (verbose and debug) were added to help us debug and see the status of the creation of the new academic years can be seen in the future. Levels of logging can be turned on and off depending on whether we are running in production vs. development. We are now also logging the number of course instances, non class events, and absences that are being created in the fall and spring semesters so that we can easily see this information as we debug in the future.