Refactor validation to show errors beneath each field
Displaying the validation beneath each field instead of at the bottom of
the modal gives greater contextual awareness to the user of precisely
what is wrong.

Some of the validation logic was also refactored to allow for multiple
things to be wrong at once. That is to say, it's possible for a field to
contain both alphabetical AND special characters (though it should be
noted that the field is checked for being numeric first, since the field
cannot be numeric AND also contain alphabetical chars)

Some tweaks were also made to the tests to swap out the `.getByText` for
`.findAllByText` as well as tweaking the code to iterate through the
validationErrors state and initialize all fields to `[]` programatically
in a couple of places. This should hopefully mean less refactoring later
if the state shape is changed to add or remove a field