Write build version into docker image
Adds an additional BUILD_VERSION arugment to our final production
Dockerfile, which will be used to write a .dockerversion file in the
/node root of our container. This can later be read in by our process to
determin which version of the app is running in the particular
container.

Adding it as a file in the image provides us with a more permanent form
of tagging, as Docker's built-in tags are mutable and can be
updated/removed at will. In contrast, any change to the BUILD_VERSION
argument will cause a new layer to be generated in the build and change
the final SHA hash; in other words, the version baked into the image
cannot be changed without publishing a new image entirely.

For our official builds on ghcr, this will use the same git version tag
that we use to tag the image and will be set by our github action.