Wrap enrollment modal fields in `<form>` element
Unfortunately, the `<button>` elements cannot currently be assocaited with the
`<form>` as mark-one does not currently appear to allow for this. A PR may
be required in mark-one to address this, but for now the input fields have been
encapsulated by a form element.

The buttons can be updated later when mark-one supports the `form` attr on
buttons