Display any server-side validation errors at the bottom of the modal
As-per the code comment - unlike the enrollment modal (which the code
for this was shamelessly ripped from) no effort is being made here to
convert the field names (i.e: "termPattern") to a human-readable format
(i.e: "Term Pattern").

This is because in the enrollment modal, these
field names are coming from the modal itself, so there is the
association of field names to human-friendly text, whereas in this
instance these field names could be literally _any_ field on the
`CourseInstance` object. This could _technically_ be brought in to the
modal component and code written to make field names human readable, but
that risks adding a lot of extra code to the component for an error
message that (hopefully) should be rarely (if ever) seen by the user.