Make buttons aria-disabled
I still feel like there's probably a better way of doing this, but this
update replaces the aria-hidden attribute on the buttons with
aria-disabled.

This should tell screenreaders that the buttons are non-functional,
while still allowing them to be tabbable for users who need keyboar
control. I'd still like to figure out a way to tell screensreaders that
the buttons are non-functional because the content they would display is
already visible to them, but I haven't found that exact solution yet.
I'll consult with Vittorio when he returns.