Create enum for HTTP Status Codes
NPM packages exist for this, however adding an npm dependency just for
what is effectively a pre-determined set of numbers seems a little
excessive.

The enum has been created in such a way as to (hopefully) be easily extensible
in the future. This includes making space for 2XX HTTP codes in the enum as an
example as to how they might fit in(along with their associated documentation)
to hopefully try and keep things neat and tidy as we go forward.