Avoid concurrent workflows
The concurrency[1] field prevents multiple instances of the same
workflow (or any two workflows with the same concurrency value) from
running at the same time. It will also cancel/replace pending workflows
when the new one attempts to run. That could put us in an edge case
where a PR test run could be skipped, but I think a skipped test run is
probably better in the long run than a failed test run. In either case,
this should not affect any active workflows.

I've added individual concurrency groups for the test, versioning,
documentation, and publication stages. The PR check workflow is simple
enough that it shouldn't have any problems running concurrently, but we
can always revisit in the future.

[1]:
https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency