Update mark-one version to 0.4.6
This is to integrate the changes made in PR [#134](https://github.com/seas-computing/mark-one/pull/134) to remove the padding from the DropListItem component. The padding was removed to fix the bug where users were clicking on the padding of each timeslot dropdown item and seeing the menu collapse without the item itself is not populated in the start and end time slots.