Add instanceId to the ScheduleView
The previous "id" field was only the id of the meeting entity, so we
need a higher-level id that we can use to match those meetings that
belong to the same course. This does, unfortunately require a migration
to add the field to the view, but fortunately does not require changing
any actual data tables.