Add fade/highlight props to blocks
First, this adds the isFaded prop to the SessionBlock, so that we can
selectively fade out the blocks that are either being hidden (via the
still-to-be-added filter) or that do not contain a meeting of the
currently chosen course.

Additionally, we're adding isFaded and isHighlighted props to the
CourseListingButton, so that we can dim out any other meetings inside of
a highlighted block when one of the courses is selected.