Set the current academic year for test purposes in retired functionality tests
These tests were previously failing, because the tests were calculating the current year as 2023, while these tests are trying to retire course instances from 2022. Since we're not allowed to retire instances of the past, the current academic year has been stubbed to 2022 so that the tests no longer fail as the real academic year changes.