Add academic year dropdown to Faculty tab
The academic year dropdown is replacing the acadYear constant that was set to 2021. Instead, the dropdown is initialized to the current academic year, which comes from the metadata context. The dropdown itself is populated with each of the existing semesters (also from the metadata context). The academicYearOptions function, which computes the years that will be displayed in the dropdown, is also used on the Courses tab, and it'll be at least partially extracted to a utils function.