Fit headings to block width
It's irritating to have the width of the rule under the column heading
change based on the length of the prefix, so we're just going to make
the heading fill the entire width of the block. This also moves the
padding that had been applied to the entire block into the heading and
list components, which fixes a slight issues with the sizing.