Apply border to entire Block
Some earlier semesters didn't have defined gaps in the schedule, so one
meeting could run 9-10:30, and the next would run 10:30-12. The meant
that two blocks would run into each other, and it could be hard to tell
where one ended and the next began. Extending the border to the bottom
of the block helps make it more clear where the separation is.