Break out Popovers from children
A consequence of the changes in the previous commit is that the
absolute/relative positioning of the scrolling wrapper "traps" the
absolute/relative positioning of the Popover, so the content is
effectively unreadable.

This moves the popovers up to the top level of the SessionBlock so that
they're no longer bound by the wrapper. This, unfortunately, requires
looping through the courses twice, but it doesn't seem to have a major
performance impact.