Add overflow indicators for scrolling content
Scrollbar styling is a browser-dependent mess, and it's hard to find a
consistent way to show a persistent scrollbar inside of an overflowing
block like this. So the (somewhat) easier solution is to dynamically
measure the positions of these blocks and their first/last entries, then
apply a custom CSS ::before/::after icon to indicate that there's more
content above and below.

I've tried to keep the indicators subtle, and adding/removing them with
CSS should avoid any extraneous detail appearing in screen readers.
However, this CSS content can't be tested with React testing library
since it only operates on the DOM.