New integration test suite
The individual controller integration tests were sufficient for testing
the integration between an API client and the controller(stubbing out
the service and repository layer below that), whilst the service
integration tests focused solely on calling a service method and then
examining the end result in the database.

Both of these were fine on their own, however they were remnants of our
previous testing strategy which falls a little short when trying to test
the interaction between the controller and the service method together.

In this case, the service method was throwing exceptions in a certain
way and in order to debug this, a suite of tests is required to refactor
and debug against. The stubbing out of different bits of the stack made
this tricky, so it made sense to write a new set of tests as API tests
but which included the database below the API also to make for a more
complete and cohesive test suite.